Added upper right window, and newScene slot
I use a Timer to update the scene being shown in my window. I got
everything working in a separate project with this implementation but it
is not working in this project for some reason. The connect between
timeout and call to update causes the program to fail if its run.   I
can't see any difference between what I have in a seperate project.
Still need to add a scroll bar that lets user control speed, and allow
user to determine the range of scenes that are looped over. That won't
take long after getting the basic concept with the timer working.  Im
going to start working on it again tomorrow and will be in the lab at
12:30.